+++ setup.cfg
@@ -33,9 +33,7 @@ packages =
 
 [entry_points]
 console_scripts =
-    aprsd = aprsd.main:main
-    aprsd-listen = aprsd.listen:main
-    aprsd-dev = aprsd.dev:main
+    aprsd = aprsd.aprsd:main
     aprsd-healthcheck = aprsd.healthcheck:main
     fake_aprs = aprsd.fake_aprs:main
